backend/requirements.txt
fastapi>=0.124.0
uvicorn>=0.38.0
# Pydantic 2.12.x pulls pydantic-core 2.41.x cp314 wheels
pydantic>=2.12.0
requests==2.32.3
starlette>=0.50.0
typing-extensions>=4.8.0
anyio==4.12.0
tzdata
PyYAML==6.0.2
python-dotenv==1.0.1
watchfiles>=0.21
python-multipart>=0.0.9

# Web scraping dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0
scrapy>=2.11.0
requests-html>=0.10.0
fake-useragent>=1.4.0
selenium>=4.15.0
webdriver-manager>=4.0.0
aiohttp>=3.9.0

# Data processing dependencies
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pymongo>=4.6.0

# Distributed task processing
celery>=5.3.0
redis>=4.6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code quality and formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# AI and ML dependencies
ddddocr>=1.4.0
easyocr>=1.7.0
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0

# AI and Machine Learning dependencies
ddddocr>=1.4.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
# 2025-12-23 自动补全说明：所有未实现但规划中的依赖已同步补全到 docs/统一接口映射表.md，补全前已自动备份 logs/backup_config_20251223T2.json。
datasets>=2.10.0
evaluate>=0.4.0

# Additional monitoring dependencies
aiofiles>=23.2.0
aiomysql>=0.1.1
aioredis>=2.0.1
